+++ docs/CHANGELOG.md
@@ -20,6 +20,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [5.10.0] - 2026-05-04
 ### Added
+- Add `cme_ids` field to Affect model for associating CME (Common Mitigation
+  Enumeration) controls with individual affects
 - Add schema validation to PURLs (OSIDB-4910)
 - Affect(s) can be automatically created and assigned to Flaw(s) for
   specific products (OSIDB-4878)